Emtpied repo in order to store twittler exercise
This repo held the Weather App which was created as part of Hack Reactor interview requirements. The app itself was designed to take advantage of multiple APIs and when user typed in a city, it gave full weather detail, exact location of the weather measurement in Google Maps, the Wikipedia summary paragraph about the city, plus the Wikipedia Main Image.

This commit text is simply a practice of git commit messages.

The repo will now contain the Twittler app, which is part of HRSF53 precourse.